Clean up Metrics/MethodLength offences
