+++ .rubocop_todo.yml
@@ -2,7 +2,7 @@ inherit_from: .rubocop.yml
 
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-02-19 13:05:41 +0000 using RuboCop version 0.64.0.
+# on 2019-02-19 14:11:20 +0000 using RuboCop version 0.64.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -19,7 +19,7 @@ Layout/IndentHeredoc:
 # Offense count: 2
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 348
+  Max: 244
 
 # Offense count: 1
 Metrics/CyclomaticComplexity:
@@ -31,17 +31,12 @@ Metrics/CyclomaticComplexity:
 Metrics/LineLength:
   Max: 137
 
-# Offense count: 6
-# Configuration parameters: CountComments, ExcludedMethods.
-Metrics/MethodLength:
-  Max: 19
-
 # Offense count: 3
 # Configuration parameters: CountKeywordArgs.
 Metrics/ParameterLists:
   Max: 8
 
-# Offense count: 2
+# Offense count: 3
 # Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist, MethodDefinitionMacros.
 # NamePrefix: is_, has_, have_
 # NamePrefixBlacklist: is_, has_, have_
@@ -70,11 +65,10 @@ Style/CaseEquality:
   Exclude:
     - 'test/test_helper.rb'
 
-# Offense count: 5
+# Offense count: 4
 # Configuration parameters: MinBodyLength.
 Style/GuardClause:
   Exclude:
-    - 'lib/bootstrap_form/form_builder.rb'
     - 'lib/bootstrap_form/helpers/bootstrap.rb'
 
 # Offense count: 1
