+++ lib/bootstrap_form/inputs/file_field.rb
@@ -12,20 +12,23 @@ module BootstrapForm
           form_group_builder(name, options) do
             content_tag(:div, class: "custom-file") do
               input_with_error(name) do
-                placeholder = options.delete(:placeholder) || "Choose file"
-                placeholder_opts = { class: "custom-file-label" }
-                placeholder_opts[:for] = options[:id] if options[:id].present?
-
-                input = file_field_without_bootstrap(name, options)
-                placeholder_label = label(name, placeholder, placeholder_opts)
-                concat(input)
-                concat(placeholder_label)
+                file_filed_input(name, options)
               end
             end
           end
         end
 
         bootstrap_alias :file_field
+      end
+
+      private
+
+      def file_filed_input(name, options)
+        placeholder = options.delete(:placeholder) || "Choose file"
+        placeholder_opts = { class: "custom-file-label" }
+        placeholder_opts[:for] = options[:id] if options[:id].present?
+
+        file_field_without_bootstrap(name, options) + label(name, placeholder, placeholder_opts)
       end
     end
   end
